requirements.txt
bcj-cffi==0.5.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
idna==2.10
multivolumefile==0.2.2
ppmd-cffi==0.3.3
psycopg2-binary==2.8.6
py7zr==0.14.1
pycparser==2.20
pycryptodome==3.10.1
pyzstd==0.14.3
requests==2.25.1
texttable==1.6.3
urllib3==1.26.4
